lyra encode / decode 関数を base.ts に集約 and ts-ignore コメントを除去
