+++ packages/sdk/src/publisher.ts
@@ -1,6 +1,6 @@
 import ConnectionBase from "./base";
 import { LYRA_MODULE } from "./base";
-import { LyraEncoder, LyraDecoder } from "@shiguredo/lyra-wasm";
+import { RTCEncodedAudioFrame } from "./types";
 
 /**
  * Role が "sendonly" または "sendrecv" の場合に Sora との WebRTC 接続を扱うクラス
@@ -99,7 +99,7 @@ export default class ConnectionPublisher extends ConnectionBase {
           if (event.track.kind == "audio") {
             const lyraDecoder = LYRA_MODULE.createDecoder({ sampleRate: 16000 });
             const transformStream = new TransformStream({
-              transform: (data, controller) => decodeFunction(lyraDecoder, data, controller),
+              transform: (data: RTCEncodedAudioFrame, controller) => this.lyraDecode(lyraDecoder, data, controller),
             });
             // eslint-disable-next-line
             receiverStreams.readable.pipeThrough(transformStream).pipeTo(receiverStreams.writable);
@@ -175,7 +175,7 @@ export default class ConnectionPublisher extends ConnectionBase {
           const senderStreams = sender.createEncodedStreams();
           if (sender.track.kind === "audio") {
             const transformStream = new TransformStream({
-              transform: (encodedFrame, controller) => encodeFunction(lyraEncoder, encodedFrame, controller),
+              transform: (data: RTCEncodedAudioFrame, controller) => this.lyraEncode(lyraEncoder, data, controller),
             });
             // eslint-disable-next-line
             senderStreams.readable.pipeThrough(transformStream).pipeTo(senderStreams.writable);
@@ -201,46 +201,4 @@ export default class ConnectionPublisher extends ConnectionBase {
     await this.waitChangeConnectionStateConnected();
     return stream;
   }
-}
-
-// @ts-ignore
-function encodeFunction(lyraEncoder: LyraEncoder, encodedFrame /*: RTCEncodedAudioFrame*/, controller) {
-  // eslint-disable-next-line
-  const view = new DataView(encodedFrame.data);
-  // eslint-disable-next-line
-  const rawData = new Float32Array(encodedFrame.data.byteLength / 2);
-  // eslint-disable-next-line
-  for (let i = 0; i < encodedFrame.data.byteLength; i += 2) {
-    const v2 = view.getInt16(i, false);
-    rawData[i / 2] = v2 / 0x7fff;
-  }
-  const encoded = lyraEncoder.encode(rawData);
-  if (encoded === undefined) {
-    // dtx
-    return;
-  }
-  // eslint-disable-next-line
-  encodedFrame.data = encoded.buffer;
-
-  // TODO: Handle DTX
-  // TODO: Reduce extra conversion between i16 and f32 (by updating lyra-wasm interface)
-
-  // eslint-disable-next-line
-  controller.enqueue(encodedFrame);
-}
-
-// @ts-ignore
-function decodeFunction(lyraDecoder: LyraDecoder, encodedFrame, controller) {
-  // TODO: handle DTX(?)
-  // eslint-disable-next-line
-  const decoded = lyraDecoder.decode(new Uint8Array(encodedFrame.data));
-  const buffer = new ArrayBuffer(decoded.length * 2);
-  const view = new DataView(buffer);
-  for (const [i, v] of decoded.entries()) {
-    view.setInt16(i * 2, v * 0x7fff, false);
-  }
-  // eslint-disable-next-line
-  encodedFrame.data = buffer;
-  // eslint-disable-next-line
-  controller.enqueue(encodedFrame);
 }
